requirements.txt
```txt
# Core
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
scipy>=1.10
pyyaml>=6.0
tqdm>=4.66

# Deep learning
torch==2.2.2            
torchvision==0.17.2

# Vision / Aug
opencv-python>=4.8
albumentations>=1.3
timm>=0.9.16

# Plots
matplotlib>=3.7
seaborn>=0.13